Move to main.cpp
